backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==4.6.0
requests==2.31.0
